+++ src/Pipedrive.net/Models/Common/Participant.cs
@@ -5,7 +5,7 @@ namespace Pipedrive
     public class Participant
     {
         [JsonProperty("person_id")]
-        public int PersonId { get; set; }
+        public long PersonId { get; set; }
 
         [JsonProperty("primary_flag")]
         public bool PrimaryFlag { get; set; }
